requirements.txt
pymongo
gevent
bottle-mongodb
bottle-cork
git+https://github.com/defnull/bottle/
beaker
webtest
pycrypto
python-magic
nose
